feat: vcの人数制限をコマンドでぱって出せるやつを追加
